Bit manipulation expressions

+++ src_smartcontract/sc_expression_arithmetic/AddExpression.cpp
@@ -53,7 +53,7 @@ AbstractVmInstance* AddExpression::interpret8Bit(VirtualMachine* vm) {
 
 	AbstractVmInstance* inst = this->list.get(0)->interpret(vm);
 	PrimitiveReference* pinst = dynamic_cast<PrimitiveReference*>(inst);
-	int8_t result = pinst->getIntValue();
+	int8_t result = pinst->getByteValue();
 
 	gc->handleFloatingObject(pinst);
 
@@ -81,7 +81,7 @@ AbstractVmInstance* AddExpression::interpret16Bit(VirtualMachine* vm) {
 
 	AbstractVmInstance* inst = this->list.get(0)->interpret(vm);
 	PrimitiveReference* pinst = dynamic_cast<PrimitiveReference*>(inst);
-	int16_t result = pinst->getIntValue();
+	int16_t result = pinst->getShortValue();
 
 	gc->handleFloatingObject(pinst);
 
@@ -136,7 +136,7 @@ AbstractVmInstance* AddExpression::interpret64Bit(VirtualMachine* vm) {
 
 	AbstractVmInstance* inst = this->list.get(0)->interpret(vm);
 	PrimitiveReference* pinst = dynamic_cast<PrimitiveReference*>(inst);
-	int64_t result = pinst->getIntValue();
+	int64_t result = pinst->getLongValue();
 
 	gc->handleFloatingObject(pinst);
